filling out cho's info to use as reference
Please use this as a reference for creating examples for other words. One example per arity (0,1,2,...) and one example using the nominalized root/compound. Keep the examples simple and clear, and remember that this is the official dictionary: no experimental grammar. I don't want to have to go back and rewrite all the examples other people have contributed, so please follow these guidelines.